Rename etcd-driver to felix-backend.
